+++ index.js
@@ -1,31 +1,6 @@
 import chalk from "chalk";
 import figlet from "figlet";
-import {
-  AddCoupon,
-  Exit,
-  GetAllCoupons,
-  GetDeliveryCosts,
-  GetPackageDetail,
-  MenuQuestion,
-  QuestionCommander,
-} from "./src/services/questions/question.service.js";
-
-async function buildQuestion() {
-  const questionCommander = new QuestionCommander();
-  questionCommander.addQuestion("default", new MenuQuestion());
-  questionCommander.addQuestion("Exit", new Exit());
-  questionCommander.addQuestion("Get Coupons", new GetAllCoupons());
-  questionCommander.addQuestion(
-    "Calculate Delivery Cost",
-    new GetDeliveryCosts()
-  );
-
-  questionCommander.addQuestion("Get Package Detail", new GetPackageDetail());
-
-  questionCommander.addQuestion("Add New Coupon", new AddCoupon());
-
-  await questionCommander.runCLIQuestion("default");
-}
+import { BuildQuestionsFacade } from "./src/services/build-question/build-questions.service.js";
 
 function setupWelcomeMessage() {
   console.log(
@@ -40,7 +15,7 @@ function setupWelcomeMessage() {
 }
 
 async function runCLI() {
-  await buildQuestion();
+  await BuildQuestionsFacade.buildQuestions();
 }
 
 setupWelcomeMessage();
