+++ src/services/questions/exit-question.service.js
@@ -1,0 +1,14 @@
+import AbstractQuestion from "./question.abstract.js";
+
+export class Exit extends AbstractQuestion {
+  constructor() {
+    super();
+  }
+
+  setOption(option) {
+    this.option = option;
+  }
+  async execute() {
+    process.exit(0);
+  }
+}
